Implement scans from both sides
Apparently, `scanRight` isn't a fair subsitution of `tails` in this
implementation. It's because of `tails` is fully non-strict (doesn't
evalutates any heads or tails of the original stream), when `scanRight`
is completely strict and evals everything like `foldRight` does (till
the end of evaluation of `=>B` in `f: (=>A, =>B)=>B`, which is never
possible in this implementation.